pyproject.toml
[tool.poetry]
name = "pilotlog"
version = "0.1.0"
description = ""
authors = ["Danylo Vasyliuk <danil.vasiliuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
psycopg = "^3.2.1"
pydantic = "^2.8.2"

[tool.poetry.group.linters.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.12.0"
mypy = "^1.11.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.5.2"
factory-boy = "^3.3.1"
pytest-cov = "^5.0.0"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
gunicorn = "^23.0.0"
gevent = "^24.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
warn_unused_configs = true

[tool.django-stubs]
django_settings_module = "pilotlot.settings"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "--reuse-db",
    "--cov",
    "--cov-report=term-missing",
    "-vv",
]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "pilotlot.settings"

[tool.coverage.run]
branch = true
relative_files = true
data_file = "/tmp/.pilotlog-service.coverage"
omit = [
    "tests",
    "**/migrations/*",
    "**/settings.py",
    "**/admin.py",
    "**/apps.py",
]
source = ["src"]


[tool.coverage.report]
skip_covered = true
show_missing = true


[tool.isort]
profile = "black"
combine_as_imports = true
lines_after_imports = 2


[tool.black]
target-version = ['py311']
